% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{color_legend}
\alias{color_legend}
\title{Legend plotting function}
\usage{
color_legend(ori.x, ori.y, colors, minmax, dim.x, dim.y, main = "", cex = 0.8)
}
\arguments{
\item{cex}{}
}
\value{

}
\description{
Legend plotting function
}
