% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_color_scale}
\alias{gs_color_scale}
\title{Color scale function generator}
\usage{
gs_color_scale(zlim, pal = "ArmyRose", pal_trim = 0)
}
\arguments{
\item{n}{}
}
\value{
get_color function
}
\description{
Color scale function generator
}
