% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_import}
\alias{gs_import}
\title{Import function for gene list data}
\usage{
gs_import(set_list, master_set)
}
\arguments{
\item{set_list}{List object of vectors containing gene names. List should be named as this is used for plotting later.}

\item{master_set}{Vector containing all gene names in the transcriptome.}
}
\value{
A list object which contains coded set information.
}
\description{
Normalizes all input gene names to match the master list and produces a base gs object
}
