% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_make_plot.df}
\alias{gs_make_plot.df}
\title{Basic plotting structure}
\usage{
gs_make_plot.df(gs, breaks, plot_type = "pval")
}
\arguments{
\item{plot_type}{}
}
\description{
Basic plotting structure
}
