% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_multi_plot}
\alias{gs_multi_plot}
\title{Multi-comparison plot}
\usage{
gs_multi_plot(
  gs,
  expand,
  value.cex = 0.8,
  label.cex = 0.8,
  zlim = NULL,
  palette = "Purple-Green",
  palette_trim = 0.2
)
}
\arguments{
\item{gs}{genesectR object}

\item{expand}{vector of set names which will be plotted along the top axis, comparing against those which are not listed.}

\item{value.cex}{scaling for values within boxes}

\item{label.cex}{scaling for axis labels}

\item{zlim}{a value used to define the maximum colors plotted. Useful for sharing the same color scale between genesects.}

\item{palette}{hcl.colors palette (found in hcl.pals()) which will be used in colorizing the genesect.}

\item{palette_trim}{a percentage of the palette which will be trimmed from the most extreme values. This is useful to remove the most dark colors (default top 0.2) from the palette, which can be difficult to read text on.}
}
\description{
Multi-comparison plot
}
\examples{

master_set <- str_glue("Gene_{1:1000}")

ls <- list(Set_A= sample(master_set, 300),
           Set_B= sample(master_set[1:100], 27),
           Set_C= sample(master_set, 99),
           Set_V= sample(master_set[1:100], 15),
           Set_W= sample(master_set, 201),
           Set_X= sample(master_set, 500),
           Set_Y= sample(master_set, 44),
           Set_Z= sample(master_set, 766))

gs <- gs_import(ls, master_set)
gs <- gs_compute_matricies(gs, mc=T)
par(mar=c(2,2,10,10))
gs_multi_plot(gs, expand=c("Set_V","Set_W","Set_X","Set_Y","Set_Z"),
              zlim=1)
}
