% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_compute_matricies}
\alias{gs_compute_matricies}
\title{Wrapper to compute all matricies}
\usage{
gs_compute_matricies(gs, mc = F)
}
\arguments{
\item{mc}{}
}
\description{
Wrapper to compute all matricies
}
