% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_multi_plot}
\alias{gs_multi_plot}
\title{Multi-comparison plot}
\usage{
gs_multi_plot(
  gs,
  expand,
  value.cex = 0.8,
  label.cex = 0.8,
  zlim = NULL,
  palette = "Purple-Green",
  palette_trim = 0.2
)
}
\arguments{
\item{label.cex}{}
}
\description{
Multi-comparison plot
}
