% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesectR.R
\name{gs_plot_fischer}
\alias{gs_plot_fischer}
\title{Wrapper to plot the fisher exact test analysis}
\usage{
gs_plot_fischer(gs, breaks = c(), gap = 0.2, margin = 5)
}
\arguments{
\item{margin}{}
}
\value{

}
\description{
Wrapper to plot the fisher exact test analysis
}
